requirements.txt
streamlit>=1.30.0
supabase>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pytest>=8.0.0
streamlit-cookies-controller>=0.0.4
plotly>=5.0.0
pandas>=2.0.0
